+++ src/main/java/domain/service/TemplateTSP.java
@@ -10,7 +10,7 @@ public abstract class TemplateTSP implements TSP {
 	
 	private Integer[] meilleureSolution;
 	protected Graphe g;
-	private int coutMeilleureSolution;
+	private double coutMeilleureSolution;
 	private int tpsLimite;
 	private long tpsDebut;
 	
@@ -34,7 +34,7 @@ public abstract class TemplateTSP implements TSP {
 		return -1;
 	}
 	
-	public int getCoutSolution(){
+	public double getCoutSolution(){
 		if (g != null)
 			return coutMeilleureSolution;
 		return -1;
@@ -47,7 +47,7 @@ public abstract class TemplateTSP implements TSP {
 	 * @return une borne inferieure du cout des chemins de <code>g</code> partant de <code>sommetCourant</code>, visitant 
 	 * tous les sommets de <code>nonVus</code> exactement une fois, puis retournant sur le sommet <code>0</code>.
 	 */
-	protected abstract int bound(Integer sommetCourant, Collection<Integer> nonVus);
+	protected abstract double bound(Integer sommetCourant, Collection<Integer> nonVus);
 	
 	/**
 	 * Methode devant etre redefinie par les sous-classes de TemplateTSP
@@ -65,7 +65,7 @@ public abstract class TemplateTSP implements TSP {
 	 * @param vus la liste des sommets deja visites (y compris sommetCrt)
 	 * @param coutVus la somme des couts des arcs du chemin passant par tous les sommets de vus dans l'ordre ou ils ont ete visites
 	 */	
-	private void branchAndBound(int sommetCrt, Collection<Integer> nonVus, Collection<Integer> vus, int coutVus){
+	private void branchAndBound(int sommetCrt, Collection<Integer> nonVus, Collection<Integer> vus, double coutVus){
 		if (System.currentTimeMillis() - tpsDebut > tpsLimite) return;
 	    if (nonVus.size() == 0){ // tous les sommets ont ete visites
 	    	if (g.estArc(sommetCrt,0)){ // on peut retourner au sommet de depart (0)
